api/requirements.txt
Flask==3.0.3
flask-sqlalchemy==3.1.1
flask_cors==5.0.0
pyodbc==5.2.0
azure-identity==1.21.0
azure-keyvault-secrets==4.9.0
weasyprint==65.1